tntdb::Value::getUnsignedLong should return an unsigned long instead of an unsigned
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@214 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/include/tntdb/bits/value.h
@@ -69,8 +69,8 @@ namespace tntdb
       long getLong() const                { return value->getLong(); }
       /// tries to convert value into an unsigned.
       unsigned getUnsigned() const        { return value->getUnsigned(); }
-      /// tries to convert value into an unsigned.
-      unsigned getUnsignedLong() const    { return value->getUnsignedLong(); }
+      /// tries to convert value into an unsigned long.
+      unsigned long getUnsignedLong() const { return value->getUnsignedLong(); }
       /// tries to convert value into an int32_t.
       int32_t getInt32() const            { return value->getInt32(); }
       /// tries to convert value into an uint32_t.
